Make Object a bootstrapfield.
And use it to define more attributes in interfaces.py

Switch from just repeating attribute names in interfaces.py to a real __all__ attribute that linters can warn about.

Change ..autoclass:: in api.rst to ..autointerface:: so we get the actual member documentation and not lots of warnings about missing __mro__.

Currently based on #53 because that reduced the dependencies of _validate_fields.

Fixes #13
