+++ src/zope/schema/_bootstrapfields.py
@@ -18,24 +18,36 @@ __docformat__ = 'restructuredtext'
 import decimal
 import fractions
 import numbers
+import threading
 from math import isinf
 
 from zope.interface import Attribute
+from zope.interface import Invalid
+from zope.interface import Interface
 from zope.interface import providedBy
 from zope.interface import implementer
+from zope.interface.interfaces import IInterface
+from zope.interface.interfaces import IMethod
 
-from zope.schema._bootstrapinterfaces import ValidationError
+from zope.event import notify
+
 from zope.schema._bootstrapinterfaces import ConstraintNotSatisfied
+from zope.schema._bootstrapinterfaces import IBeforeObjectAssignedEvent
 from zope.schema._bootstrapinterfaces import IContextAwareDefaultFactory
 from zope.schema._bootstrapinterfaces import IFromUnicode
+from zope.schema._bootstrapinterfaces import IValidatable
 from zope.schema._bootstrapinterfaces import NotAContainer
 from zope.schema._bootstrapinterfaces import NotAnIterator
 from zope.schema._bootstrapinterfaces import RequiredMissing
+from zope.schema._bootstrapinterfaces import SchemaNotCorrectlyImplemented
+from zope.schema._bootstrapinterfaces import SchemaNotFullyImplemented
+from zope.schema._bootstrapinterfaces import SchemaNotProvided
 from zope.schema._bootstrapinterfaces import StopValidation
 from zope.schema._bootstrapinterfaces import TooBig
 from zope.schema._bootstrapinterfaces import TooLong
 from zope.schema._bootstrapinterfaces import TooShort
 from zope.schema._bootstrapinterfaces import TooSmall
+from zope.schema._bootstrapinterfaces import ValidationError
 from zope.schema._bootstrapinterfaces import WrongType
 
 from zope.schema._compat import text_type
@@ -655,3 +667,167 @@ class Int(Integral):
     """
     _type = integer_types
     _unicode_converters = (int,)
+
+
+VALIDATED_VALUES = threading.local()
+
+def _validate_fields(schema, value):
+    errors = {}
+    # Interface can be used as schema property for Object fields that plan to
+    # hold values of any type.
+    # Because Interface does not include any Attribute, it is obviously not
+    # worth looping on its methods and filter them all out.
+    if schema is Interface:
+        return errors
+    # if `value` is part of a cyclic graph, we need to break the cycle to avoid
+    # infinite recursion. Collect validated objects in a thread local dict by
+    # it's python represenation. A previous version was setting a volatile
+    # attribute which didn't work with security proxy
+    if id(value) in VALIDATED_VALUES.__dict__:
+        return errors
+    VALIDATED_VALUES.__dict__[id(value)] = True
+    # (If we have gotten here, we know that `value` provides an interface
+    # other than zope.interface.Interface;
+    # iow, we can rely on the fact that it is an instance
+    # that supports attribute assignment.)
+
+    try:
+        for name in schema.names(all=True):
+            attribute = schema[name]
+            if IMethod.providedBy(attribute):
+                continue # pragma: no cover
+
+            try:
+                if IValidatable.providedBy(attribute):
+                    # validate attributes that are fields
+                    attribute.validate(getattr(value, name))
+                # XXX: We're not even checking the existence of non-IField
+                # Attribute objects.
+            except ValidationError as error:
+                errors[name] = error
+            except AttributeError as error:
+                # property for the given name is not implemented
+                errors[name] = SchemaNotFullyImplemented(error).with_field_and_value(attribute, None)
+    finally:
+        del VALIDATED_VALUES.__dict__[id(value)]
+    return errors
+
+
+class _BoundSchema(object):
+    """
+    This class proxies a schema to get its fields bound to a context.
+    """
+
+    __slots__ = ('schema', 'context')
+
+    def __new__(cls, schema, context):
+        # Only proxy if we really need to.
+        if schema is Interface or context is None:
+            return schema
+        return object.__new__(cls)
+
+    def __init__(self, schema, context):
+        self.schema = schema
+        self.context = context
+
+    def __getitem__(self, name):
+        # Indexing this item will bind fields,
+        # if possible
+        attr = self.schema[name]
+        try:
+            return attr.bind(self.context)
+        except AttributeError:
+            return attr
+
+    # but let all the rest slip to schema
+    def __getattr__(self, name):
+        return getattr(self.schema, name)
+
+
+class Object(Field):
+    """
+    Implementation of :class:`zope.schema.interfaces.IObject`.
+    """
+    schema = None
+
+    def __init__(self, schema=_NotGiven, **kw):
+        """
+        Object(schema=<Not Given>, *, validate_invariants=True, **kwargs)
+
+        Create an `~.IObject` field. The keyword arguments are as for `~.Field`.
+
+        .. versionchanged:: 4.6.0
+           Add the keyword argument *validate_invariants*. When true (the default),
+           the schema's ``validateInvariants`` method will be invoked to check
+           the ``@invariant`` properties of the schema.
+        .. versionchanged:: 4.6.0
+           The *schema* argument can be ommitted in a subclass
+           that specifies a ``schema`` attribute.
+        """
+        if schema is _NotGiven:
+            schema = self.schema
+
+        if not IInterface.providedBy(schema):
+            raise WrongType
+
+        self.schema = schema
+        self.validate_invariants = kw.pop('validate_invariants', True)
+        super(Object, self).__init__(**kw)
+
+    def _validate(self, value):
+        super(Object, self)._validate(value)
+
+        # schema has to be provided by value
+        if not self.schema.providedBy(value):
+            raise SchemaNotProvided(self.schema, value).with_field_and_value(self, value)
+
+        # check the value against schema
+        schema_error_dict = _validate_fields(_BoundSchema(self.schema, value), value)
+        invariant_errors = []
+        if self.validate_invariants:
+            try:
+                self.schema.validateInvariants(value, invariant_errors)
+            except Invalid:
+                # validateInvariants raises a wrapper error around
+                # all the errors it got if it got errors, in addition
+                # to appending them to the errors list. We don't want
+                # that, we raise our own error.
+                pass
+
+        if schema_error_dict or invariant_errors:
+            errors = list(schema_error_dict.values()) + invariant_errors
+            exception = SchemaNotCorrectlyImplemented(
+                errors,
+                self.__name__
+            ).with_field_and_value(self, value)
+            exception.schema_errors = schema_error_dict
+            exception.invariant_errors = invariant_errors
+            try:
+                raise exception
+            finally:
+                # Break cycles
+                del exception
+                del invariant_errors
+                del schema_error_dict
+                del errors
+
+    def set(self, object, value):
+        # Announce that we're going to assign the value to the object.
+        # Motivation: Widgets typically like to take care of policy-specific
+        # actions, like establishing location.
+        event = BeforeObjectAssignedEvent(value, self.__name__, object)
+        notify(event)
+        # The event subscribers are allowed to replace the object, thus we need
+        # to replace our previous value.
+        value = event.object
+        super(Object, self).set(object, value)
+
+
+@implementer(IBeforeObjectAssignedEvent)
+class BeforeObjectAssignedEvent(object):
+    """An object is going to be assigned to an attribute on another object."""
+
+    def __init__(self, object, name, context):
+        self.object = object
+        self.name = name
+        self.context = context
